+++ CHANGES.md
@@ -10,6 +10,7 @@ Version 3.1.0
 - fixed filtering of the damage prevention
 - made breakblock and placeblock filterable
 - made the item prevention ignore blocks (item ID <= 256)
+- fixed enabling preventions that are disabled in the config
 
 Version 3.0.3
 -------------